ref(core): Split `performance` module
Split code related to header parsing into its own module.

This change is primarily just moving stuff around. But, one test has been modified to not need access to private APIs that are no longer in the same module after the split.
